Registration, login, JWT authentication tokens and authorization

+++ FlightAttendant/Controllers/AirportsController.cs
@@ -1,7 +1,11 @@
-﻿using FlightAttendant.Data.Dtos.Airports;
+﻿using FlightAttendant.Auth.Model;
+using FlightAttendant.Data.Dtos.Airports;
 using FlightAttendant.Data.Entities;
 using FlightAttendant.Data.Repositories;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.JsonWebTokens;
+using System.Security.Claims;
 
 namespace FlightAttendant.Controllers
 {
@@ -10,9 +14,11 @@ namespace FlightAttendant.Controllers
     public class AirportsController : ControllerBase
     {
         private readonly IAirportsRepository _airportsRepository;
-        public AirportsController(IAirportsRepository airportsRepository)
+        private readonly IAuthorizationService _authorizationService;
+        public AirportsController(IAirportsRepository airportsRepository, IAuthorizationService authorizationService)
         {
             _airportsRepository = airportsRepository;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet]
@@ -38,10 +44,12 @@ namespace FlightAttendant.Controllers
         }
 
         [HttpPost]
+        [Authorize(Roles = FlightAttendantRoles.AttendantUser)]
         public async Task<ActionResult<AirportDto>> Create(CreateAirportDto createAirportDto)
         {
-            var airport = new Airport {Name = createAirportDto.Name, Location = createAirportDto.Location};
+            var airport = new Airport {Name = createAirportDto.Name, Location = createAirportDto.Location, UserId = User.FindFirstValue(JwtRegisteredClaimNames.Sub)};
             await _airportsRepository.CreateAsync(airport);
+
 
             //201
             return Created($"/api/airports/{airport.Id}/", new AirportDto(airport.Id, airport.Name, airport.Location));
@@ -50,6 +58,7 @@ namespace FlightAttendant.Controllers
 
         [HttpPut]
         [Route("{airportId}")]
+        [Authorize(Roles = FlightAttendantRoles.AttendantUser)]
         public async Task<ActionResult<AirportDto>> Update(int airportId, UpdateAirportDto updateAirportDto)
         {
             var airport = await _airportsRepository.GetOneAsync(airportId);
@@ -58,6 +67,13 @@ namespace FlightAttendant.Controllers
             if (airport == null)
             {
                 return NotFound($"Couldn't find am airport with id of {airportId}");
+            }
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, airport, PolicyNames.ResourceOwner);
+
+            if (!authorizationResult.Succeeded)
+            {
+                return NotFound("");
             }
 
             airport.Name = updateAirportDto.Name;
@@ -69,6 +85,7 @@ namespace FlightAttendant.Controllers
 
         [HttpDelete]
         [Route("{airportId}")]
+        [Authorize(Roles = FlightAttendantRoles.AttendantUser)]
         public async Task<ActionResult> Remove(int airportId)
         {
             var airport = await _airportsRepository.GetOneAsync(airportId);
@@ -76,7 +93,14 @@ namespace FlightAttendant.Controllers
             //404
             if (airport == null)
             {
-                return NotFound($"Couldn't find am airport with id of {airportId}");
+                return NotFound($"Couldn't find an airport with id of {airportId}");
+            }
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, airport, PolicyNames.ResourceOwner);
+
+            if (!authorizationResult.Succeeded)
+            {
+                return NotFound("");
             }
 
             await _airportsRepository.RemoveAsync(airport);
